+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Allow setting passing a `Sequel::Model` to `:account` in `Rodauth::Rails.rodauth` (@janko)
+
 ## 0.17.1 (2021-10-20)
 
 * Skip checking CSRF when request forgery protection wasn't loaded on the controller (@janko)